Added address label as Flow component attribute.
This allows the user to set the address label to be shown as title below the search field, but above the address detail fields. This is especially useful if 'Detailed Address Fields Required' is set to true, as it will add a label where otherwise only a red asteriks is shown, indicating that the whole inputAddress component is set to required.